Refactor clone command (#505)
